parking_backend/requirements.txt
alembic==1.16.2
aniso8601==10.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1
black==23.11.0
flake8==7.0.0
pytest==7.4.4
coverage==7.3.2
requests==2.31.0
prometheus-flask-exporter==2.0.0
greenlet==3.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
